requirements.txt
numpy
pandas
dotenv
seaborn
plotly
matplotlib
optuna
plotly
scikit-learn
xgboost 
joblib
google-genai
requests
fastapi
pydantic
uvicorn